+++ packages/react-reconciler/src/ReactFiberLane.js
@@ -7,30 +7,10 @@
  * @flow
  */
 
-import type {FiberRoot, ReactPriorityLevel} from './ReactInternalTypes';
+import type {FiberRoot, ReactPriorityLevel, Lane, LanePriority, LaneMap, Lanes} from './ReactInternalTypes';
 
-export opaque type LanePriority =
-  | 0
-  | 1
-  | 2
-  | 3
-  | 4
-  | 5
-  | 6
-  | 7
-  | 8
-  | 9
-  | 10
-  | 11
-  | 12
-  | 13
-  | 14
-  | 15
-  | 16
-  | 17;
-export opaque type Lanes = number;
-export opaque type Lane = number;
-export opaque type LaneMap<T> = Array<T>;
+// unwind-cycles: Re-export lane definitions here
+export type {Lane, LanePriority, LaneMap, Lanes};
 
 import invariant from 'shared/invariant';
 
@@ -41,7 +21,7 @@ import {
   LowPriority as LowSchedulerPriority,
   IdlePriority as IdleSchedulerPriority,
   NoPriority as NoSchedulerPriority,
-} from './SchedulerWithReactIntegration.new';
+} from './ReactFiberSchedulerPriorities';
 
 export const SyncLanePriority: LanePriority = 15;
 export const SyncBatchedLanePriority: LanePriority = 14;
