+++ packages/react-reconciler/src/ReactInternalTypes.js
@@ -20,13 +20,13 @@ import type {SuspenseInstance} from './ReactFiberHostConfig';
 import type {WorkTag} from './ReactWorkTags';
 import type {TypeOfMode} from './ReactTypeOfMode';
 import type {Flags} from './ReactFiberFlags';
-import type {Lane, LanePriority, Lanes, LaneMap} from './ReactFiberLane';
 import type {RootTag} from './ReactRootTags';
 import type {TimeoutHandle, NoTimeout} from './ReactFiberHostConfig';
 import type {Wakeable} from 'shared/ReactTypes';
 import type {Interaction} from 'scheduler/src/Tracing';
 
-// Unwind Circular: moved from ReactFiberHooks.old
+export type {ReactPriorityLevel} from './ReactFiberSchedulerPriorities';
+
 export type HookType =
   | 'useState'
   | 'useReducer'
@@ -43,7 +43,28 @@ export type HookType =
   | 'useMutableSource'
   | 'useOpaqueIdentifier';
 
-export type ReactPriorityLevel = 99 | 98 | 97 | 96 | 95 | 90;
+export type LanePriority =
+  | 0
+  | 1
+  | 2
+  | 3
+  | 4
+  | 5
+  | 6
+  | 7
+  | 8
+  | 9
+  | 10
+  | 11
+  | 12
+  | 13
+  | 14
+  | 15
+  | 16
+  | 17;
+export type Lanes = number;
+export type Lane = number;
+export type LaneMap<T> = Array<T>;
 
 export type ContextDependency<T> = {
   context: ReactContext<T>,
