+++ packages/react-reconciler/src/SchedulerWithReactIntegration.old.js
@@ -8,6 +8,14 @@
  */
 
 import type {ReactPriorityLevel} from './ReactInternalTypes';
+import {
+  ImmediatePriority,
+  UserBlockingPriority,
+  NormalPriority,
+  LowPriority,
+  IdlePriority,
+  NoPriority,
+} from './ReactFiberSchedulerPriorities';
 
 // Intentionally not named imports because Rollup would use dynamic dispatch for
 // CommonJS interop named imports.
@@ -58,17 +66,6 @@ export type SchedulerCallback = (isSync: boolean) => SchedulerCallback | null;
 type SchedulerCallbackOptions = {timeout?: number, ...};
 
 const fakeCallbackNode = {};
-
-// Except for NoPriority, these correspond to Scheduler priorities. We use
-// ascending numbers so we can compare them like numbers. They start at 90 to
-// avoid clashing with Scheduler's priorities.
-export const ImmediatePriority: ReactPriorityLevel = 99;
-export const UserBlockingPriority: ReactPriorityLevel = 98;
-export const NormalPriority: ReactPriorityLevel = 97;
-export const LowPriority: ReactPriorityLevel = 96;
-export const IdlePriority: ReactPriorityLevel = 95;
-// NoPriority is the absence of priority. Also React-only.
-export const NoPriority: ReactPriorityLevel = 90;
 
 export const shouldYield = Scheduler_shouldYield;
 export const requestPaint =
